Fix what appears to be the last cbob glitch.
This glitch was occuring because the additional USB interrupts from the camera were occasionally causing our timer irq to be called much later, running a transaction for the cbob late. Because the next timer IRQ would be called on time, the next transaction would occur without the 1.5ms gap from the previous transaction, which on occasion caused desync. Now we restart the timer every time the transaction is run, so we're guaranteed 1.5ms or more between transactions.